reseting App.js
